requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
apispec==3.3.1
apistar==0.7.2
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docopt==0.6.2
entrypoints==0.3
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==2.0.0
h11==0.9.0
httptools==0.0.13
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
marshmallow==3.7.0
matplotlib==3.2.2
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.19.0
packaging==20.4
pandocfilters==1.4.2
parso==0.7.0
pdf2image==1.13.1
pdfminer.six==20200517
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pkg-resources==0.0.0
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycryptodome==3.9.8
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-multipart==0.0.5
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
requests-toolbelt==0.9.1
responder==2.0.5
rfc3986==1.4.0
Rx==1.6.1
Send2Trash==1.5.0
six==1.15.0
sortedcontainers==2.2.2
starlette==0.12.13
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typesystem==0.2.4
urllib3==1.25.9
uvicorn==0.10.9
uvloop==0.14.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
whitenoise==5.1.0
widgetsnbextension==3.5.1
zipp==3.1.0
